iNT-7559 - create command to generate data sources taht will be used to create a file integrationDataSources.json
