+++ packages/integration-sdk-cli/src/commands/index.ts
@@ -10,3 +10,4 @@ export * from './neo4j';
 export * from './visualize-dependencies';
 export * from './generate-integration-graph-schema';
 export * from './troubleshoot';
+export * from './generate-integration-data-sources';
